requirements.txt
asgiref==3.7.2
Django==4.2.5
django-auditlog==2.3.0
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.4.4
packaging==23.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
